+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-09 12:50:03
+
+### Changed Files
+- `supabase/functions/gmail-send-email/index.ts`
+
+### Statistics
+```
+ 1 file changed, 39 insertions(+), 14 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-09 11:31:53
 
 ### Changed Files